+++ package.json
@@ -6,9 +6,9 @@
   "license": "agpl",
   "private": true,
   "scripts": {
-    "dev": "webpack --config js-src/webpack.dev.js",
-    "watch": "webpack --progress --watch --config js-src/webpack.dev.js",
-    "build": "webpack --progress --hide-modules --config js-src/webpack.prod.js"
+    "dev": "webpack --config src/webpack.dev.js",
+    "watch": "webpack --progress --watch --config src/webpack.dev.js",
+    "build": "webpack --progress --hide-modules --config src/webpack.prod.js"
   },
   "dependencies": {
     "vue": "^2.5.16"
