+++ src/richObjectStringParser.js
