+++ src/webpack.common.js
